pyproject.toml
[project]
name = "multi-agent-inspection-system"
version = "0.1.0"
description = "A multi-agent tunnel inspection system built around OpenAI Structured Outputs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.54.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

